% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constrain.R
\name{bound.sets}
\alias{bound.sets}
\title{Bounds sets.}
\usage{
bound.sets(constrained)
}
\arguments{
\item{constrained}{a constrained edge list (see
\code{\link{community.sampler}})}
}
\value{
Returns sets of weights bounded above by their first
  element, in topological order.
}
\description{
Construct bounding sets used to order weights to meet the edge
weight constraints.
}
\details{
These bounds sets are used to order a set of random edge weights
so that they meet the imposed constraints.
}
