% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tk.r
\name{checkbox}
\alias{checkbox}
\alias{checkcolumn}
\title{Checkbox widgets}
\usage{
checkbox(parent, label, initial = 0)

checkcolumn(parent, label, rows, label.rows = T)
}
\arguments{
\item{parent}{the parent window.}

\item{label}{the label for the enclosing frame.}

\item{initial}{the initial state of the checkbox}

\item{rows}{the row labels}

\item{label.rows}{whether to label rows.}
}
\value{
Returns an object of class checkbox or checkcolumn with elements
\item{\code{window}}{the widget}
\item{\code{selected}}{function that returns the state of the checkboxes}
\item{\code{state}}{the tclVars representing the state of the checkboxes}
}
\description{
Construct a checkbox widget
}
\details{
The \code{checkbox} function makes a single checkbox widget, while
\code{checkcolumn} makes a widget containing a column of
checkboxes.
}

