% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constrain.R
\name{constraint.order}
\alias{constraint.order}
\title{Constraint Sorting}
\usage{
constraint.order(w, bounds)
}
\arguments{
\item{w}{absolute values of the edge weights}

\item{bounds}{a bounds set generated by \code{\link{bound.sets}}}
}
\value{
The absolute weights ordered to meet the edge weights
  constraints represented by the bounds set.
}
\description{
Order a set of absolute edge weights to meet imposed edge weights
constraints.
}
