% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Community.R
\name{drop.nodes}
\alias{drop.nodes}
\title{Drop one or more nodes from a system}
\usage{
drop.nodes(sim, to.drop, method = "remove")
}
\arguments{
\item{sim}{the result from \code{system.simulate}}

\item{to.drop}{the names of the nodes to drop (check \code{node.labels(sim$edges)})}

\item{method}{either "remove" (the specified nodes will be fully removed from the
system) or "zeros" (the specified nodes will be left in the system but all edges
from or to these nodes (other than self-interactions) are set to zero).}
}
\value{
As for \code{system.simulate}
}
\description{
This is an experimental function! Given a set of system simulation outputs
(from \code{system.simulate}), it will drop one or more nodes and their
associated edges, but leave all other elements of the system untouched. Each
set of edge weights in \code{sim} is checked for stability after dropping
the specified nodes, and any matrices representing unstable systems are
removed from the returned set.
}
