% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{grviz.digraph}
\alias{grviz.digraph}
\title{Export to DOT}
\usage{
grviz.digraph(edges, name = "web", fontsize = 10, node.style = "filled",
  node.shape = "oval", node.color = "DarkOrange", edge.color = "DarkGrey")
}
\arguments{
\item{edges}{An edge list}

\item{name}{The name of the digraph}

\item{fontsize}{Fontsize for node labels.}

\item{node.style}{The node style.}

\item{node.shape}{The node shape.}

\item{node.color}{The node color.}

\item{edge.color}{The edge color.}
}
\value{
Returns a string.
}
\description{
Write a DOT specification of the model.
}
\details{
Write a DOT specification of the model in a form suitable for use
with \code{grViz} from \pkg{DiagrammeR}.
}

