% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{impact.barplot.shiny}
\alias{impact.barplot.shiny}
\title{Shiny Impact Barplot}
\usage{
impact.barplot.shiny(sim, epsilon = 1e-05, main = "", cex.axis = 1)
}
\arguments{
\item{sim}{the result from \code{system.simulate}}

\item{epsilon}{outomes below this in absolute magnitude are treated as zero.}

\item{main}{text for plot title}

\item{cex.axis}{character expansion factor for the edge labels}
}
\description{
A Shiny app to display the impact of a perturbation as a barplot
}
\details{
This control constructs a barplot that shows the fraction of
simulations in which a positive (orange), negative (blue) or zero
(off white) outcome occurs at each node following a given
perturbation.

The user may specify the perturbation of the nodes, and any
outcome known from monitoring the network, and then construct a
barplot of the frequency table of outcomes at each node.
}
\examples{
\dontrun{
  set.seed(32)
  ## Sample model
  edges <- parse.digraph(c(
    "E *-> D",
    "D *-> C",
    "C -> E",
    "E *-> B",
    "B *-> A",
    "A -> E",
    "D --> B"))
  edges <- enforce.limitation(edges)
  sims <- system.simulate(10, edges)
  impact.barplot.shiny(sims)
}

}
