% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk.r
\name{impact.table}
\alias{impact.table}
\title{Impact Table}
\usage{
impact.table(sim, epsilon = 1e-05)
}
\arguments{
\item{sim}{the result from \code{system.simulate}}

\item{epsilon}{outomes below this in absolute magnitude are treated as zero.}
}
\value{
The crosstabulation as a matrix
}
\description{
Tabulate the impact of every positive perturbation as table.
}
\details{
Crosstabulate the mean impact (positive or negative) at each node
for a positive perturbation of each node.  The k-th column
corresponds to a perturbation of the k-th node, and shows the mean
impact on each node.
}
