% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constrain.R
\name{parse.constraints}
\alias{parse.constraints}
\title{Text Representations of Edge Constraints}
\usage{
parse.constraints(lines, edges)
}
\arguments{
\item{lines}{a vector of strings representing "a < b" constraints}

\item{edges}{an edge list}
}
\value{
Returns a constrained edge list - a list with components
\item{\code{edges}}{the edge list}
\item{\code{a}}{the index of the "a" edge of the constraint}
\item{\  code{b}}{the index of the "b" edge of the constraint}
}
\description{
Read text representations of edge weight constraints
}
\details{
Read a set of edge constraints from a vector of strings,
represented in the following format.  Each string corresponds to a
constraint of the form "a < b", and where a and b are two edges.
Each edge consists of two node labels separated by an arrow, where
the arrow consists of a sequence of dashes "-" followed by one of
the character sequences ">","*","<>". The number of dashes used in
the arrow is ignored.
}
