% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{parse.edge}
\alias{parse.edge}
\title{Indices of (Directed) Edges}
\usage{
parse.edge(lines, edges)
}
\arguments{
\item{lines}{a vector of strings representing directed edges}

\item{edges}{an edge list}
}
\value{
the indices of the directed edges within the edge list
}
\description{
Parse a text representation of (directed) edges, return the index
of the directed edge within the edge list.
}
\details{
Each directed edge is represented as a string consisting of two
node labels separated by an arrow, where the arrow consists of a
sequence of dashes "-" followed by one of the character sequences
">","*","<>". The number of dashes used in the arrow is ignored.
}
\examples{
## Sample model
edges <- parse.digraph(c(
  "E *-> D",
  "D *-> C",
  "C -> E",
  "E *-> B",
  "B *-> A",
  "A -> E",
  "D --> B"))
edges <- enforce.limitation(edges)
parse.edge(c("E->D","D-*E","A-*B"),edges)
}

