% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Community.R, R/XML.R
\docType{package}
\name{QPress-package}
\alias{QPress-package}
\title{Qualitative Network Analysis}
\description{
Qualitiative analysis of press perturbations of network models.
Given a network model represented as a signed directed graphs, this
package provide facilities for evaluating the impact of a press
perturbation to the system through simulation.

This package provides facilities for simulating press
perturbation scenarios for qualitative network models
specified as signed directed graphs (signed digraphs).
}
\author{
Ben Raymond, Jessica Melbourne-Thomas, Simon Wotherspoon

B. Raymond, J. Melbourne-Thomas and S. Wotherspoon
}
