% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Community.R
\name{rescale.predators}
\alias{rescale.predators}
\title{Predator Scaling}
\usage{
rescale.predators(edges, scaling = 0.2)
}
\arguments{
\item{edges}{an edge list}

\item{scaling}{the scaling to apply}
}
\value{
Returns an edge list augmented with scaling information on predator-prey edges.
}
\description{
Apply a relative scaling to weights in predator-prey relationships
}
\details{
The per-capita effect of a prey on its predator should in theory be
smaller than that of the predator on its prey.
This function adds a scaling to each predator-prey pair such that
abs(a_ij) <= abs(scaling*a_ji), where i is the predator and j is the prey
}

