% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk.r
\name{stability.density}
\alias{stability.density}
\alias{stability.density0}
\title{Stability Density Plots}
\usage{
stability.density(sim, main = "")

stability.density0(sim, edges, smooth = 1, main = "")
}
\arguments{
\item{sim}{the result from \code{system.simulate0}}

\item{main}{text for plot title}

\item{edges}{logical vector indicating which edges to plot.}

\item{smooth}{double in the range [0,1] controlling the level of smoothing applied.}
}
\description{
Display weights of valid and invalid matrices as a density plots
}
\details{
This control constructs density plots that show the distribution
of selected edge weights for the cases that meet the selected
validation criteria (blue), and those that do not (red), following
a given perturbation.

The slider controls the level of smoothing of the densities.
Edges are labelled by pairs of integers for compactness, where the
integer codes correspond to the ordering of the node labels.

\code{stability.density0} is a non-interactive variant for
programmatic use.
}
